+++ src/components/pages/contactpage.tsx
@@ -65,7 +65,7 @@ const AboutPage: React.FC = () => {
 		<div className="h-screen w-screen flex flex-col relative">
 			<div className="relative z-10 flex flex-col h-full w-full">
 				<Header />
-				<h1 className="text-3xl font-bold m-4 flex justify-center item-center">
+				<h1 className="text-3xl font-bold m-4 flex justify-center item-center overflow-y-auto pt-14">
 					Contact Us
 				</h1>
 				<div className="flex flex-1 w-full">